+++ shared/src/simulation/activities/index.ts
@@ -8,3 +8,4 @@ export * from './load-vehicle';
 export * from './send-remote-event';
 export * from './transfer-vehicle';
 export * from './publish-radiogram';
+export * from './transfer-patient-to-hospital';